+++ .idea/workspace.xml
@@ -20,11 +20,11 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="7c551163-f241-4a1b-a5c0-5d116641d997" name="Default Changelist" comment="Updated">
-      <change afterPath="$PROJECT_DIR$/src/main/java/ConnectionToDB/ConnectionPull.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/java/ConnectionToDB/DBConnection.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/resources/dbconfiguration.properties" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/pom.xml" beforeDir="false" afterPath="$PROJECT_DIR$/pom.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/ConnectionToDB/ConnectionPull.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/ConnectionToDB/ConnectionPoll.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/ConnectionToDB/DBConnection.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/ConnectionToDB/DBConnection.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/resources/dbconfiguration.properties" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/dbconfiguration.properties" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -66,9 +66,9 @@
       <recent name="C:\JavaTasks\Kh041Java_ORM_Team1" />
     </key>
   </component>
-  <component name="RunManager" selected="Application.DBConnection">
-    <configuration name="ConnectionPull" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
-      <option name="MAIN_CLASS_NAME" value="ConnectionToDB.ConnectionPull" />
+  <component name="RunManager" selected="Application.ConnectionPoll">
+    <configuration name="ConnectionPoll" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
+      <option name="MAIN_CLASS_NAME" value="ConnectionToDB.ConnectionPoll" />
       <module name="Kh041Java_ORM_Team1" />
       <extension name="coverage">
         <pattern>
@@ -95,8 +95,8 @@
     </configuration>
     <recent_temporary>
       <list>
+        <item itemvalue="Application.ConnectionPoll" />
         <item itemvalue="Application.DBConnection" />
-        <item itemvalue="Application.ConnectionPull" />
       </list>
     </recent_temporary>
   </component>
@@ -113,7 +113,7 @@
       <workItem from="1587580894532" duration="120000" />
       <workItem from="1587581856923" duration="1290000" />
       <workItem from="1587630760823" duration="4260000" />
-      <workItem from="1587669814488" duration="7998000" />
+      <workItem from="1587669814488" duration="18233000" />
     </task>
     <servers />
   </component>
@@ -141,22 +141,22 @@
       <screen x="0" y="0" width="1536" height="834" />
     </state>
     <state x="549" y="172" key="FileChooserDialogImpl/0.0.1536.834@0.0.1536.834" timestamp="1587670350785" />
-    <state width="1493" height="213" key="GridCell.Tab.0.bottom" timestamp="1587678918581">
+    <state width="1493" height="145" key="GridCell.Tab.0.bottom" timestamp="1587830306930">
       <screen x="0" y="0" width="1536" height="834" />
     </state>
-    <state width="1493" height="213" key="GridCell.Tab.0.bottom/0.0.1536.834@0.0.1536.834" timestamp="1587678918581" />
-    <state width="1493" height="213" key="GridCell.Tab.0.center" timestamp="1587678918581">
+    <state width="1493" height="145" key="GridCell.Tab.0.bottom/0.0.1536.834@0.0.1536.834" timestamp="1587830306930" />
+    <state width="1493" height="145" key="GridCell.Tab.0.center" timestamp="1587830306930">
       <screen x="0" y="0" width="1536" height="834" />
     </state>
-    <state width="1493" height="213" key="GridCell.Tab.0.center/0.0.1536.834@0.0.1536.834" timestamp="1587678918581" />
-    <state width="1493" height="213" key="GridCell.Tab.0.left" timestamp="1587678918581">
+    <state width="1493" height="145" key="GridCell.Tab.0.center/0.0.1536.834@0.0.1536.834" timestamp="1587830306930" />
+    <state width="1493" height="145" key="GridCell.Tab.0.left" timestamp="1587830306930">
       <screen x="0" y="0" width="1536" height="834" />
     </state>
-    <state width="1493" height="213" key="GridCell.Tab.0.left/0.0.1536.834@0.0.1536.834" timestamp="1587678918581" />
-    <state width="1493" height="213" key="GridCell.Tab.0.right" timestamp="1587678918581">
+    <state width="1493" height="145" key="GridCell.Tab.0.left/0.0.1536.834@0.0.1536.834" timestamp="1587830306930" />
+    <state width="1493" height="145" key="GridCell.Tab.0.right" timestamp="1587830306930">
       <screen x="0" y="0" width="1536" height="834" />
     </state>
-    <state width="1493" height="213" key="GridCell.Tab.0.right/0.0.1536.834@0.0.1536.834" timestamp="1587678918581" />
+    <state width="1493" height="145" key="GridCell.Tab.0.right/0.0.1536.834@0.0.1536.834" timestamp="1587830306930" />
     <state x="249" y="0" key="SettingsEditor" timestamp="1587669870755">
       <screen x="0" y="0" width="1536" height="834" />
     </state>
