+++ pom.xml
@@ -33,6 +33,7 @@
         <artifactId>mysql-connector-java</artifactId>
         <version>8.0.19</version>
     </dependency>
+
     </dependencies>
 
 </project>
