+++ src/main/resources/dbconfiguration.properties
@@ -1,4 +1,4 @@
 URL=jdbc:postgresql://localhost:5432/hibernatepresentation
-USER=mysqluser
-PASS=Thisispasword1!
+USER=hibernameuser
+PASS=Thisispassword
 DRIVER=org.postgresql.Driver
